+++ Dyadminoes/Match.m
@@ -120,7 +120,7 @@
   while (dataDyadminoIndexesThisTurn.count < kNumDyadminoesInRack && self.pile.count > 0) {
     NSLog(@"dataDyadminoIndicesThisTurn is %luu",(unsigned long)dataDyadminoIndexesThisTurn.count);
     NSUInteger randIndex = [self randomIntegerUpTo:self.pile.count];
-    NSLog(@"randomIndex is %i", randIndex);
+    NSLog(@"randomIndex is %lu", (unsigned long)randIndex);
     DataDyadmino *dataDyad = self.pile[randIndex];
       // rack order is total count at the time
     dataDyad.myRackOrder = [NSNumber numberWithInteger:dataDyadminoIndexesThisTurn.count];
